requirements.txt
deepagents 
tavily-python
dotenv
langchain_mistralai
langchain_community
arxiv
pymupdf
fitz
streamlit 